Update Addon Communication Protocol
This changes the previous communication protocol to instead use a separate server for addon<->server communication.

The reason being is because it would require rewriting the EO packet encoders and decoders in order to use the same socket since sending multiple messages in a row causes issues to occur due to chunking.