Use a default area of the entire map.
The intended effect should be as follows:
"Timed based events such as When countdown timer # goes off wouldn't do anything with lines that have the triggering player, unless you used a line to make someone specifically targeted become the triggering player. Anything with *any player present* would affect everyone on the map, or if given areas and filters, would affect a group of people in those places."

This is related to issue #5.